requirements.txt
pyyaml
python-telegram-bot
anthropic==0.3.4
GoogleBard==2.1.0
